Move the serve env parsing into serve/config.rs
